Removed luceneindexmanager and routed flow through indexservice
