--- Conceptpower+Spring/src/main/java/edu/asu/conceptpower/servlet/lucene/ILuceneIndexManger.java
@@ -1,11 +1,0 @@
-package edu.asu.conceptpower.servlet.lucene;
-
-import edu.asu.conceptpower.servlet.exceptions.LuceneException;
-
-public interface ILuceneIndexManger {
-
-    public void deleteIndexes() throws LuceneException;
-
-    public void indexConcepts() throws LuceneException, IllegalArgumentException, IllegalAccessException ;
-
-}
